Build 233
Added: All items involved in combination or separation, are now required to be fully repaired.

Added: If the player's inventory is full when items are returned, they're dropped on the ground below the player.

Added: All open HavenElytra GUIs open as the server is reloaded or shut down, will close and return all items.

NOTE:
HavenElytra is one of my first plugins, and therefore outdated code wise.
If I am to do more or better updates, I'll have to rewrite the plugin to use newer methods, and the latest update of ValorlessUtils.
The plugin still works as is though.